initial v0.2 commit includes changes required to integrate the new tectonic model with previous algoritms for momentum and erosion this commit should produce similar results to the initial github commit

+++ WorldGenerator/Grid.cpp
@@ -26,7 +26,7 @@ namespace WorldBuilder {
         }
     }
     
-    Grid::Grid(size_t vertexCount) : verts(vertexCount) {
+    Grid::Grid(uint32_t vertexCount) : verts(vertexCount) {
         
     }
